# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Astronomy Observation API",
    "description": "API for managing users and astronomy observations",
    "version": "1.0.0"
  },
  "host": "project-two-hu6p.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/observations/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}